+++ packages/tko.computed/spec/proxyBehaviour.js
@@ -1,6 +1,6 @@
 
 import {
-  proxy, unproxy, peek, isProxied, getObservable
+  proxy, peek, isProxied, getObservable
 } from '../src/proxy'
 
 import {
@@ -26,15 +26,28 @@ describe('Proxy', function () {
     p.a = 124
     expect(p.a).toBe(124)
     expect(x.a).toBe(124)
+    // Add a new property
     p.b = 456
     expect(x.b).toBe(456)
     expect(p.b).toBe(456)
     expect(peek(p, 'b')).toBe(456)
   })
 
-  it('unproxies to the original', function () {
+  it('unproxies to the original by calling the proxy with no args', function () {
     const x = {}
-    expect(unproxy(proxy(x))).toEqual(x)
+    expect(proxy(x)()).toEqual(x)
+  })
+
+  it('assigns own properties when the proxy is called with an object', function () {
+    const x = { a: 9, b: 4 }
+    const p = proxy(x)
+    p({ b: 5, c: 12 })
+    expect(p.a).toBe(9)
+    expect(p.b).toBe(5)
+    expect(p.c).toBe(12)
+    expect(x.a).toBe(9)
+    expect(x.b).toBe(5)
+    expect(x.c).toBe(12)
   })
 
   it('creates dependencies on the proxied elements', function () {
@@ -46,14 +59,18 @@ describe('Proxy', function () {
   })
 
   it('does not create dependencies with `peek`', function () {
-    const p = proxy({ a: 123 })
+    const p = proxy({ a: 123, b () { return this.a + 7 } })
     let pa = 0
+    let pb = 0
     computed(() => {
       pa = peek(p, 'a')
+      pb = peek(p, 'b')
     })
     expect(pa).toBe(123)
+    expect(pb).toBe(130)
     p.a++
     expect(pa).toBe(123)
+    expect(pb).toBe(130)
   })
 
   it('exposes the observable with getObservable', function () {
@@ -71,5 +88,11 @@ describe('Proxy', function () {
     p.b = 8
     expect(x()).toBe(8)
     expect(p.a).toBe(16)
+  })
+
+  it('allow adding computeds', function () {
+    const p = proxy({ x: 4 })
+    p.x2 = function () { return this.x * this.x }
+    expect(p.x2).toBe(16)
   })
 })